fix: logger branch cleanups (#1226)
