+++ codegen/src/shuttle_main/mod.rs
@@ -256,7 +256,6 @@ impl ToTokens for Loader {
             async fn loader(
                 mut #factory_ident: shuttle_runtime::ProvisionerFactory,
                 mut #resource_tracker_ident: shuttle_runtime::ResourceTracker,
-                deployment_id: String,
             ) -> #return_type {
                 use shuttle_runtime::Context;
                 #extra_imports
@@ -326,7 +325,6 @@ mod tests {
             async fn loader(
                 mut _factory: shuttle_runtime::ProvisionerFactory,
                 mut _resource_tracker: shuttle_runtime::ResourceTracker,
-                deployment_id: String,
             ) -> ShuttleSimple {
                 use shuttle_runtime::Context;
                 simple().await
@@ -395,7 +393,6 @@ mod tests {
             async fn loader(
                 mut factory: shuttle_runtime::ProvisionerFactory,
                 mut resource_tracker: shuttle_runtime::ResourceTracker,
-                deployment_id: String,
             ) -> ShuttleComplex {
                 use shuttle_runtime::Context;
                 use shuttle_runtime::{Factory, ResourceBuilder};
@@ -507,7 +504,6 @@ mod tests {
             async fn loader(
                 mut factory: shuttle_runtime::ProvisionerFactory,
                 mut resource_tracker: shuttle_runtime::ResourceTracker,
-                deployment_id: String,
             ) -> ShuttleComplex {
                 use shuttle_runtime::Context;
                 use shuttle_runtime::{Factory, ResourceBuilder};
