+++ common/src/backends/tracing.rs
@@ -13,7 +13,6 @@ use opentelemetry::{
 };
 use opentelemetry_http::HeaderExtractor;
 use opentelemetry_otlp::WithExportConfig;
-use opentelemetry_proto::tonic::common::v1::{any_value, AnyValue};
 use pin_project::pin_project;
 use tower::{Layer, Service};
 use tracing::{debug_span, instrument::Instrumented, Instrument, Span, Subscriber};
@@ -145,97 +144,4 @@ where
 
         ExtractPropagationFuture { response_future }
     }
-}
-
-fn serde_json_value_to_any_value(
-    value: serde_json::Value,
-) -> Option<opentelemetry_proto::tonic::common::v1::AnyValue> {
-    use opentelemetry_proto::tonic::common::v1::any_value::Value;
-
-    let value = match value {
-        serde_json::Value::Null => return None,
-        serde_json::Value::Bool(b) => Value::BoolValue(b),
-        serde_json::Value::Number(n) => {
-            if n.is_f64() {
-                Value::DoubleValue(n.as_f64().unwrap()) // Safe to unwrap as we just checked if it is a f64
-            } else {
-                Value::IntValue(n.as_i64().unwrap()) // Safe to unwrap since we know it is not a f64
-            }
-        }
-        serde_json::Value::String(s) => Value::StringValue(s),
-        serde_json::Value::Array(a) => {
-            Value::ArrayValue(opentelemetry_proto::tonic::common::v1::ArrayValue {
-                values: a
-                    .into_iter()
-                    .flat_map(serde_json_value_to_any_value)
-                    .collect(),
-            })
-        }
-        serde_json::Value::Object(o) => {
-            Value::KvlistValue(opentelemetry_proto::tonic::common::v1::KeyValueList {
-                values: serde_json_map_to_key_value_list(o),
-            })
-        }
-    };
-
-    Some(opentelemetry_proto::tonic::common::v1::AnyValue { value: Some(value) })
-}
-
-/// Convert a [serde_json::Map] into an anyvalue [KeyValue] list
-pub fn serde_json_map_to_key_value_list(
-    map: serde_json::Map<String, serde_json::Value>,
-) -> Vec<opentelemetry_proto::tonic::common::v1::KeyValue> {
-    map.into_iter()
-        .map(
-            |(key, value)| opentelemetry_proto::tonic::common::v1::KeyValue {
-                key,
-                value: serde_json_value_to_any_value(value),
-            },
-        )
-        .collect()
-}
-
-/// Convert an [AnyValue] to a [serde_json::Value]
-pub fn from_any_value_to_serde_json_value(any_value: AnyValue) -> serde_json::Value {
-    let Some(value) = any_value.value else {
-        return serde_json::Value::Null;
-    };
-
-    match value {
-        any_value::Value::StringValue(s) => serde_json::Value::String(s),
-        any_value::Value::BoolValue(b) => serde_json::Value::Bool(b),
-        any_value::Value::IntValue(i) => serde_json::Value::Number(i.into()),
-        any_value::Value::DoubleValue(f) => {
-            let Some(number) = serde_json::Number::from_f64(f) else {
-                return serde_json::Value::Null;
-            };
-            serde_json::Value::Number(number)
-        }
-        any_value::Value::ArrayValue(a) => {
-            let values = a
-                .values
-                .into_iter()
-                .map(from_any_value_to_serde_json_value)
-                .collect();
-
-            serde_json::Value::Array(values)
-        }
-        any_value::Value::KvlistValue(kv) => {
-            let map = from_any_value_kv_to_serde_json_map(kv.values);
-
-            serde_json::Value::Object(map)
-        }
-        any_value::Value::BytesValue(_) => serde_json::Value::Null,
-    }
-}
-
-/// Convert a [KeyValue] list in a [serde_json::Map]
-pub fn from_any_value_kv_to_serde_json_map(
-    kv_list: Vec<opentelemetry_proto::tonic::common::v1::KeyValue>,
-) -> serde_json::Map<String, serde_json::Value> {
-    let iter = kv_list
-        .into_iter()
-        .flat_map(|kv| Some((kv.key, from_any_value_to_serde_json_value(kv.value?))));
-
-    serde_json::Map::from_iter(iter)
 }
