+++ logger/Cargo.toml
@@ -6,11 +6,9 @@ license.workspace = true
 repository.workspace = true
 
 [dependencies]
-async-broadcast = "0.5.1"
 async-trait = { workspace = true }
 chrono = { workspace = true }
 clap = { workspace = true }
-opentelemetry-proto = { workspace = true, features = ["gen-tonic", "traces"] }
 prost-types = { workspace = true }
 serde_json = { workspace = true }
 sqlx = { workspace = true, features = [
