+++ runtime/src/alpha/mod.rs
@@ -183,14 +183,13 @@ where
         self,
         factory: Fac,
         resource_tracker: ResourceTracker,
-        deployment_id: String,
     ) -> Result<Self::Service, shuttle_service::Error>;
 }
 
 #[async_trait]
 impl<F, O, Fac, S> Loader<Fac> for F
 where
-    F: FnOnce(Fac, ResourceTracker, String) -> O + Send,
+    F: FnOnce(Fac, ResourceTracker) -> O + Send,
     O: Future<Output = Result<S, shuttle_service::Error>> + Send,
     Fac: Factory + 'static,
     S: Service,
@@ -201,9 +200,8 @@ where
         self,
         factory: Fac,
         resource_tracker: ResourceTracker,
-        deployment_id: String,
     ) -> Result<Self::Service, shuttle_service::Error> {
-        (self)(factory, resource_tracker, deployment_id).await
+        (self)(factory, resource_tracker).await
     }
 }
 
@@ -221,7 +219,6 @@ where
             resources,
             secrets,
             service_name,
-            deployment_id,
         } = request.into_inner();
         println!("loading alpha service at {path}");
 
@@ -263,63 +260,62 @@ where
         let loader = self.loader.lock().unwrap().deref_mut().take().unwrap();
 
         // send to new thread to catch panics
-        let service =
-            match tokio::spawn(loader.load(factory, resource_tracker, deployment_id)).await {
-                Ok(res) => match res {
-                    Ok(service) => service,
-                    Err(error) => {
-                        println!("loading service failed: {error}");
-
-                        let message = LoadResponse {
-                            success: false,
-                            message: error.to_string(),
-                            resources: new_resources
-                                .lock()
-                                .expect("to get lock no new resources")
-                                .iter()
-                                .map(resource::Response::to_bytes)
-                                .collect(),
-                        };
-                        return Ok(Response::new(message));
-                    }
-                },
+        let service = match tokio::spawn(loader.load(factory, resource_tracker)).await {
+            Ok(res) => match res {
+                Ok(service) => service,
                 Err(error) => {
-                    let resources = new_resources
-                        .lock()
-                        .expect("to get lock no new resources")
-                        .iter()
-                        .map(resource::Response::to_bytes)
-                        .collect();
+                    println!("loading service failed: {error}");
 
-                    if error.is_panic() {
-                        let panic = error.into_panic();
-                        let msg = match panic.downcast_ref::<String>() {
-                            Some(msg) => msg.to_string(),
-                            None => match panic.downcast_ref::<&str>() {
-                                Some(msg) => msg.to_string(),
-                                None => "<no panic message>".to_string(),
-                            },
-                        };
-
-                        println!("loading service panicked: {msg}");
-
-                        let message = LoadResponse {
-                            success: false,
-                            message: msg,
-                            resources,
-                        };
-                        return Ok(Response::new(message));
-                    } else {
-                        println!("loading service crashed: {error}");
-                        let message = LoadResponse {
-                            success: false,
-                            message: error.to_string(),
-                            resources,
-                        };
-                        return Ok(Response::new(message));
-                    }
+                    let message = LoadResponse {
+                        success: false,
+                        message: error.to_string(),
+                        resources: new_resources
+                            .lock()
+                            .expect("to get lock no new resources")
+                            .iter()
+                            .map(resource::Response::to_bytes)
+                            .collect(),
+                    };
+                    return Ok(Response::new(message));
                 }
-            };
+            },
+            Err(error) => {
+                let resources = new_resources
+                    .lock()
+                    .expect("to get lock no new resources")
+                    .iter()
+                    .map(resource::Response::to_bytes)
+                    .collect();
+
+                if error.is_panic() {
+                    let panic = error.into_panic();
+                    let msg = match panic.downcast_ref::<String>() {
+                        Some(msg) => msg.to_string(),
+                        None => match panic.downcast_ref::<&str>() {
+                            Some(msg) => msg.to_string(),
+                            None => "<no panic message>".to_string(),
+                        },
+                    };
+
+                    println!("loading service panicked: {msg}");
+
+                    let message = LoadResponse {
+                        success: false,
+                        message: msg,
+                        resources,
+                    };
+                    return Ok(Response::new(message));
+                } else {
+                    println!("loading service crashed: {error}");
+                    let message = LoadResponse {
+                        success: false,
+                        message: error.to_string(),
+                        resources,
+                    };
+                    return Ok(Response::new(message));
+                }
+            }
+        };
 
         *self.service.lock().unwrap() = Some(service);
